Refactor parser scope analysis
